Flip tree levels
Occasionally, the wordle_tree will split at a position, then a subtree will immediately split again at the same position. This partitions the search space into 3 distinct groups, which only differ by the possible values in one position.

e.g. {'a', 'b', 'c'} -> {'a'} and {'b', 'c'} -> {'a'} and {'b'} and {'c'}

The same partition can be made in a variety of ways. Because the encoding depends on the number of valid_words in a tree and the number of possible values in the tree_split_position, these encodings can vary in efficiency.

This code searches for these inefficiencies and fixes them when found.