Use number of possible word combos instead of bits
When encoding the words of a leaf node, we already find the index of the combination of words in a theoretical list. However, we need to store this info. Previously, we had taken (the ceiling of the log2 of the max possible value) number of bits, but now we just multiply by the max possible value. Savings can range from 0 to 1 bits per tree.